Fix linker PDB output.  Move some PDB generation code to dxcutil.
